move SettingsDto to shared objects
