+++ pom.xml
@@ -47,6 +47,11 @@
 			<version>3.1</version>
 		</dependency>
 		<dependency>
+			<groupId>pl.smarthouse</groupId>
+			<artifactId>shared-objects</artifactId>
+			<version>1.1</version>
+		</dependency>
+		<dependency>
 			<groupId>org.springdoc</groupId>
 			<artifactId>springdoc-openapi-webflux-ui</artifactId>
 			<version>1.6.9</version>
